\name{lm.restricted}
\alias{lm.restricted}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Quadratic method in second step of estimate.
}
\description{
Quadratic optimization method is used to ensure that the estimate for betas are non-negative.
}
\usage{
lm.restricted(y, X)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{
Response vector.
}
  \item{X}{
Design Matrix.
}
}

\value{
The estimate of betas are returned
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}


