\name{SNP.Pattern}
\alias{SNP.Pattern}
\docType{data}
\title{
Data for SNP Pattern
}
\description{
The data describes the SNP Pattern.
}
\usage{data(SNP.Pattern)}
\format{
  A data frame with 15387 observations. Numbers from 1 to 14 represents the order of the reference genomes used to generate the SNP pattern file
}
\examples{
data(SNP.Pattern)
}
